requirements.txt
langgraph
langchain
langchain-openai
langchain-astradb
langchain-community
cassio
python-dotenv
tiktoken
python-multipart
fastapi
uvicorn
pydantic
numpy
astrapy
openai
IPython
langchain-cohere
cohere
PyMuPDF
chardet
pinecone-client
rank-bm25
langchain-experimental
nltk
spacy